# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hakimasrori24@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@mailinator.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Test",
                                    "phone": "0987654321",
                                    "email": "test@mailinator.com",
                                    "password": "password",
                                    "confirm_password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Me",
                "description": "Me",
                "operationId": "Me",
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/category": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "CategoryShow",
                "description": "CategoryShow",
                "operationId": "CategoryShow",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "integer"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "page": {
                                        "type": "integer"
                                    },
                                    "per_page": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "search": "",
                                    "type": 0,
                                    "category_id": 0,
                                    "slug": "",
                                    "page": 1,
                                    "per_page": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                }
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "ProductShow",
                "description": "ProductShow",
                "operationId": "ProductShow",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "search": "",
                                    "outlet_id": 0,
                                    "product_id": 0,
                                    "slug": "",
                                    "tags": "",
                                    "page": 1,
                                    "per_page": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/product/store": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "ProductStore",
                "description": "ProductStore",
                "operationId": "ProductStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "categories",
                                    "name",
                                    "price",
                                    "description",
                                    "images"
                                ],
                                "properties": {
                                    "images": {
                                        "description": "Array of product images (max 5)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "enable_variant": {
                                        "description": "Indicates if variants are enabled",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "variants": {
                                        "description": "Array of product variants (required if enable_variant=true)",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "price",
                                                "status"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "description": "Variant name",
                                                    "type": "string",
                                                    "maxLength": 100,
                                                    "example": "Size M"
                                                },
                                                "price": {
                                                    "description": "Variant price",
                                                    "type": "number",
                                                    "format": "float",
                                                    "minimum": 0,
                                                    "example": 20
                                                },
                                                "status": {
                                                    "description": "Variant status",
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS/ERROR by code in json result"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Category",
            "description": "Category"
        },
        {
            "name": "Product",
            "description": "Product"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}